The support for native int64 fail, because for example, without explicit cast, (-5 : uint32) is equal to (0xfffffffb : int64) instead of (0xfffffffffffffffb : int64).
git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/simsoc-cert@1553 ce709078-17db-4df8-89f4-d3e5bb8270d9

+++ arm6/simlight/slv6_iss_c_prelude.h
@@ -22,5 +22,5 @@ static bool not_cpy_instr(uint32_t bincode) {
 }
 
 static int32_t to_int32(uint32_t x) {return (int32_t)x;}
-static int64 to_i64(uint32_t x) {return I64_of_int32(x);}
+static int64 to_i64(uint32_t x) {return I64_of_int32((int32_t)x);}
 static uint64 to_u64(uint32_t x) {return I64_lsr(I64_lsl(I64_of_int32(x), 32), 32);}
